+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Version 0.5.0
 
+- Core: Implemented SaveData functions
+- CRT: Added the save functions: `vgs_save`, `vgs_load` and `vgs_save_check`
 - CRT: Added the i-math inline functions: `vgs_abs` and `vgs_sgn`
 - CRT: Added a string function: `vgs_stricmp`
 - CRT: Added a ctype function: `vgs_atoi`